Move kafka producer to a more generic directory (#2712)
