+++ udata/event/__init__.py
@@ -1,0 +1,8 @@
+from enum import Enum
+from .producer import produce, KafkaProducerSingleton # noqa
+
+
+class KafkaMessageType(Enum):
+    INDEX = 'index'
+    REINDEX = 'reindex'
+    UNINDEX = 'unindex'
